user-module-gles: Remove the gles kernel module dependencies
The existing user-module-gles recipe has a dependency on kernel-module-gles.
This dependency causes the UI toolkit to be rebuilt in the event of a
Linux Kernel change.  Because kernel-module-gles has a build dependency
on the Linux kernel.  As a result, the UI toolkit starts rebuilding,
which is caused by the OpenGLES/EGL dependency.

Signed-off-by: Naoto Yamaguchi <naoto.yamaguchi@aisin.co.jp>